Split translation suggestion into separate controller
